Logger has been moved and implemented as an internal lib
