+++ apps/backend/src/app.module.ts
@@ -4,7 +4,7 @@ import { ThrottlerModule } from '@nestjs/throttler';
 import { AuthModule } from './auth/auth.module';
 import { AppCacheModule } from './cache/cache.module';
 import { AppConfigModule, AppConfigService } from '@reduced.to/config';
-import { AppLoggerModule } from './logger/logger.module';
+import { AppLoggerModule } from '@reduced.to/logger';
 import { NovuModule } from './novu/novu.module';
 import { PrismaService } from '@reduced.to/prisma';
 import { UniqueConstraint } from './shared/decorators/unique/unique.decorator';
