+++ libs/logger/README.md
@@ -1,0 +1,7 @@
+# logger
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test logger` to execute the unit tests via [Jest](https://jestjs.io).
