+++ libs/logger/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/logger.module';
+export * from './lib/logger.service';
